# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/tables": {
            "get": {
                "tags": [
                    "Tables"
                ],
                "summary": "List all tables",
                "operationId": "a86641f3dcf512e2bed1751c36c1ea8c",
                "responses": {
                    "200": {
                        "description": "List of tables",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tables"
                ],
                "summary": "Create a new table",
                "operationId": "6e292739794946df4226d64299dd14ab",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "data"
                                ],
                                "properties": {
                                    "data": {
                                        "properties": {
                                            "attributes": {
                                                "required": [
                                                    "name",
                                                    "capacity"
                                                ],
                                                "properties": {
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "capacity": {
                                                        "type": "integer"
                                                    },
                                                    "location": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Table created successfully"
                    }
                }
            }
        },
        "/api/v1/tables/{id}": {
            "get": {
                "tags": [
                    "Tables"
                ],
                "summary": "Get table by ID",
                "operationId": "057ecb07a05a3dbc0b40a425635a7270",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Table data"
                    },
                    "404": {
                        "description": "Table not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Tables"
                ],
                "summary": "Update a table",
                "operationId": "ecae6d497fc1d508d2363457cc465cea",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "data"
                                ],
                                "properties": {
                                    "data": {
                                        "properties": {
                                            "attributes": {
                                                "properties": {
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "capacity": {
                                                        "type": "integer"
                                                    },
                                                    "location": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Table updated successfully"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tables"
                ],
                "summary": "Delete a table",
                "operationId": "73c62001661c1cd75c30085f921f518c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Table deleted successfully"
                    }
                }
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}